+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-# 0.1.0
-- moved the special implementation of writes to any polarity register to the Expander Trait implementation instead of overwriting StandardExpanderInterface Trait functions
+# 1.1.0
+- Added `Debug` trait implementation for all types which are accessible by the library user
+- Added `Clone` and `Copy` trait implementation for `Polarity` enum
+- Internal code cleanup
 
 # 1.0.0
 **Breaking changes!**
 - Refactored error types and generics. The whole error handling is now simpler and should make more sense in general, as underlying embedded-hal errors are directly passed to the `ExpanderError` enum. Due to those changes certain types need an additional generic for I2C interface. Generics have been refactored as well so some generics are not on the same position like they used to be. Migrating to 1.0 should be relatively simple by adding those missing generics or rearranging them.
-- Added `std::error::Error` trait implementation for `ExapnderError`. This is automatically enabled by using the crates `std` feature. The change should allow for easier error handling with existing std solutions and libraries.
+- Added `std::error::Error` trait implementation for `ExpanderError`. This is automatically enabled by using the crates `std` feature. The change should allow for easier error handling with existing std solutions and libraries.
 - Updated crate to rust 2021 edition
+
+# 0.1.0
+- moved the special implementation of writes to any polarity register to the Expander Trait implementation instead of overwriting StandardExpanderInterface Trait functions